+++ src/main/java/com/gdg/z_meet/domain/meeting/entity/Team.java
@@ -1,0 +1,34 @@
+package com.gdg.z_meet.domain.meeting.entity;
+
+import com.gdg.z_meet.global.common.BaseEntity;
+import com.gdg.z_meet.domain.user.entity.enums.Gender;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class Team extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "team_id", unique = true)
+    private Long id;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private TeamType teamType;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private Gender gender;
+
+    @Column(nullable = false)
+    @Builder.Default
+    private Integer hi = 2;
+}
