+++ src/main/java/com/gdg/z_meet/domain/user/entity/User.java
@@ -1,0 +1,27 @@
+package com.gdg.z_meet.domain.user.entity;
+
+import com.gdg.z_meet.global.common.BaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class User extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "user_id", unique = true)
+    private Long id;
+
+    @Column(unique = true, nullable = false)
+    private String name;
+
+    @Column(unique = true, nullable = false)
+    private String studentNumber;
+
+    @Column(nullable = false)
+    private String password;
+}
