+++ src/main/java/com/gdg/z_meet/domain/user/entity/enums/Emoji.java
@@ -1,0 +1,4 @@
+package com.gdg.z_meet.domain.user.entity.enums;
+
+public enum Emoji {
+}
