+++ src/main/java/com/gdg/z_meet/domain/user/entity/enums/Major.java
@@ -1,0 +1,38 @@
+package com.gdg.z_meet.domain.user.entity.enums;
+
+public enum Major {
+    //인문계열
+    KOREAN, PHILOSOPHY, KOREANHISTORY,
+    //어문계열
+    ENGLISH, CHINESE, JAPANESE, FRENCH,
+    //음악과
+    MUSIC,
+    //신학과
+    THEOLOGY,
+
+    SOCIALWELFARE, PSYCHOLOGY, SOCIOLOGY, SPECIAL_EDUCATION,
+    //경영계열
+    BUSINESS, ACCOUNTING,
+    //국제, 법정경계열
+    INTERNATIONAL_STUDIES, LAW, ECONOMICS, PUBLIC_ADMINISTRATION,
+    //글로벌경영대학
+    GBS,
+
+    //자연과학계열
+    CHEMISTRY, MATH, PHYSICS,
+    //생명과학계열
+    DESIGN, CLOTHING, CHILDREN, FOOD_NUTRITION,
+    //의생명과학과
+    MEDICAL_BIOLOGY,
+    //약학대학
+    PHARMACY,
+    //간호대, 의대
+    NURSING, MEDICINE,
+
+    //ICT공학계열
+    CSIE, MTC, ICE,
+    //바이오융합공학계열
+    BIOTECH, ENVI, BMCE,
+    //인공지능, 데사, 바이오메디컬소프트웨어
+    AI, DA, BMSW
+}
