+++ src/main/java/com/gdg/z_meet/domain/user/entity/enums/Style.java
@@ -1,0 +1,6 @@
+package com.gdg.z_meet.domain.user.entity.enums;
+
+public enum Style {
+    CASUAL, STREET, NEAT, VINTAGE,
+    HEAP, SPORTY, FORMAL, CUTIE
+}
