--- src/main/java/com/gdg/z_meet/entity/Notification.java
@@ -1,4 +1,0 @@
-package com.gdg.z_meet.entity;
-
-public class Notification {
-}
