--- src/main/java/com/gdg/z_meet/entity/Token.java
@@ -1,26 +1,0 @@
-package com.gdg.z_meet.entity;
-
-import com.gdg.z_meet.entity.common.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Token extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "token_id", unique = true)
-    private Long id;
-
-    private String accessToken;
-
-    private String refreshToken;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-}
