--- src/main/java/com/gdg/z_meet/entity/TossPayments.java
@@ -1,22 +1,0 @@
-package com.gdg.z_meet.entity;
-
-import com.gdg.z_meet.entity.common.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class TossPayments extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "toss_payments_id", unique = true)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "order_id")
-    private Order order;
-}
