--- src/main/java/com/gdg/z_meet/entity/common/BaseEntity.java
@@ -1,25 +1,0 @@
-package com.gdg.z_meet.entity.common;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-@Getter
-public abstract class BaseEntity {
-
-    @CreatedDate
-    @JsonFormat(timezone = "Asia/Seoul")
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    @JsonFormat(timezone = "Asia/Seoul")
-    private LocalDateTime updatedAt;
-}
