--- src/main/java/com/gdg/z_meet/entity/enums/Category.java
@@ -1,5 +1,0 @@
-package com.gdg.z_meet.entity.enums;
-
-public enum Category {
-    DATE, FOOD, GOODS, EVENT
-}
