--- src/main/java/com/gdg/z_meet/entity/enums/Emoji.java
@@ -1,4 +1,0 @@
-package com.gdg.z_meet.entity.enums;
-
-public enum Emoji {
-}
