--- src/main/java/com/gdg/z_meet/entity/enums/Grade.java
@@ -1,5 +1,0 @@
-package com.gdg.z_meet.entity.enums;
-
-public enum Grade {
-    FIRST, SECOND, THIRD, FOURTH
-}
