--- src/main/java/com/gdg/z_meet/entity/enums/IdealAge.java
@@ -1,5 +1,0 @@
-package com.gdg.z_meet.entity.enums;
-
-public enum IdealAge {
-    OLDER, YOUNGER, SAME, NO_MATTER
-}
