--- src/main/java/com/gdg/z_meet/entity/enums/IdealType.java
@@ -1,6 +1,0 @@
-package com.gdg.z_meet.entity.enums;
-
-public enum IdealType {
-    DOG, CAT, HAMSTER, FOX,
-    BEAR, DINOSAUR, RABBIT, WOLF
-}
