--- src/main/java/com/gdg/z_meet/entity/enums/Music.java
@@ -1,6 +1,0 @@
-package com.gdg.z_meet.entity.enums;
-
-public enum Music {
-    KPOP, POP, BALLOD, HIPHOP,
-    JPOP, ZAZZ, BAND, CLASSICAL
-}
