--- src/main/java/com/gdg/z_meet/entity/enums/Style.java
@@ -1,6 +1,0 @@
-package com.gdg.z_meet.entity.enums;
-
-public enum Style {
-    CASUAL, STREET, NEAT, VINTAGE,
-    HEAP, SPORTY, FORMAL, CUTIE
-}
