--- src/main/java/com/gdg/z_meet/entity/enums/TeamType.java
@@ -1,5 +1,0 @@
-package com.gdg.z_meet.entity.enums;
-
-public enum TeamType {
-    TWO_TO_TWO, THREE_TO_THREE
-}
