--- src/main/java/com/gdg/z_meet/entity/mapping/JoinChat.java
@@ -1,28 +1,0 @@
-package com.gdg.z_meet.entity.mapping;
-
-import com.gdg.z_meet.entity.ChatRoom;
-import com.gdg.z_meet.entity.User;
-import com.gdg.z_meet.entity.common.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class JoinChat extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "join_chat_id", unique = true)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "chat_room_id")
-    private ChatRoom chatRoom;
-}
