--- src/main/java/com/gdg/z_meet/entity/mapping/UserTeam.java
@@ -1,28 +1,0 @@
-package com.gdg.z_meet.entity.mapping;
-
-import com.gdg.z_meet.entity.Team;
-import com.gdg.z_meet.entity.User;
-import com.gdg.z_meet.entity.common.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class UserTeam extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_team_id", unique = true)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "team_id")
-    private Team team;
-}
