+++ src/main/java/com/gdg/z_meet/global/common/Notification.java
@@ -1,0 +1,4 @@
+package com.gdg.z_meet.global.common;
+
+public class Notification {
+}
